ath79: ar71xx create a separate driver for ar71xx pci interrupt controller.
It is based on Chuanhong Guo work.

PCI interrupt controller is not part of PCI. It is a part of reset controller
with 0x18060018, 0x1806001c control registers.

This should fix a bug with one IRQ for all PCI devices and also handle the PCI_CORE interrupt.

I am not sure that this kind of cascading is good for performance, but this way we are more flexible when cofiguring IRQs.

Run tested on DIR-825 B1.